moving mutex.cs to shared
